1.24 release

+++ libnf/perl/Net-NfDump/testdist.sh
@@ -1,11 +1,11 @@
 #!/bin/bash 
 
-set -x
+#set -x
 
 export COPYFILE_DISABLE=1
 
-make dist 
-make dist 
+echo "y" | make dist 
+echo "Y" | make dist 
 
 HOSTS="
 	root@hawk.cis.vutbr.cz 
@@ -19,17 +19,21 @@ HOSTS="
 	root@test-openbsd.net.vutbr.cz
 "
 
-DIST="Net-NfDump-1.23"
+DIST="Net-NfDump-1.24"
 EXT=".tar.gz"
 
 for h in $HOSTS; do 
 	echo "***********************************************"
 	echo ${h}
 	echo "***********************************************"
-	scp ${DIST}${EXT} ${h}:/tmp/
-	ssh ${h} "export AUTOMATED_TESTING=1 ; cd /tmp/ && (rm -rf ${DIST}/ ; tar xzf ${DIST}${EXT}) && cd ${DIST} && perl Makefile.PL && make -j 8 -j 8 && make test ; rm -rf Net-Dump-*"
-	echo FINISHED: ${h}
-	echo 
+	echo "Press Enter to start test or 0 to skip..."
+	read a 
+	if [ "$a" != "0" ]; then
+		scp ${DIST}${EXT} ${h}:/tmp/
+		ssh ${h} "export AUTOMATED_TESTING=1 ; cd /tmp/ && (rm -rf ${DIST}/ ; tar xzf ${DIST}${EXT}) && cd ${DIST} && perl Makefile.PL && make -j 8 -j 8 && make test ; rm -rf Net-Dump-*"
+		echo FINISHED: ${h}
+		echo 
+	fi
 
 done
